# [PERF] textparse: lightweight `p.isCreatedSeries()`