@Maniktherana 2024-10-11
Okay, found a much faster one
can barely see `p.Next` in the memory profile its a tiny sliver


<img width="615" alt="Screenshot 2024-10-11 at 10 39 19 PM" src="https://github.com/user-attachments/assets/28045d82-3f66-4fcd-b85a-87b8c2b1f2e7">

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/textparse
cpu: Apple M1
                              │   main.txt   │            this-pr.txt             │           pr-15150v2.txt           │
                              │    sec/op    │    sec/op     vs base              │   sec/op     vs base               │
Parse omtestdata.txt omtext     35.55µ ± 11%   33.20µ ± 22%       ~ (p=0.310 n=6)   30.14µ ± 5%  -15.22% (p=0.002 n=6)
Parse promtestdata.txt omtext   2.573m ±  1%   2.510m ±  2%  -2.44% (p=0.026 n=6)   2.566m ± 8%        ~ (p=0.937 n=6)
geomean                         302.4µ         288.7µ        -4.54%                 278.1µ        -8.04%

                              │   main.txt    │             this-pr.txt             │           pr-15150v2.txt            │
                              │      B/s      │      B/s       vs base              │     B/s       vs base               │
Parse omtestdata.txt omtext     120.1Mi ± 10%   128.7Mi ± 18%       ~ (p=0.310 n=6)   141.7Mi ± 5%  +17.94% (p=0.002 n=6)
Parse promtestdata.txt omtext   12.36Mi ±  1%   12.68Mi ±  2%  +2.55% (p=0.024 n=6)   12.40Mi ± 7%        ~ (p=0.985 n=6)
geomean                         38.54Mi         40.40Mi        +4.84%                 41.91Mi        +8.75%

                              │   main.txt   │             this-pr.txt             │           pr-15150v2.txt            │
                              │     B/op     │     B/op      vs base               │     B/op      vs base               │
Parse omtestdata.txt omtext     17.44Ki ± 0%   13.88Ki ± 0%  -20.43% (p=0.002 n=6)   10.55Ki ± 0%  -39.52% (p=0.002 n=6)
Parse promtestdata.txt omtext   164.2Ki ± 0%   133.4Ki ± 0%  -18.79% (p=0.002 n=6)   108.4Ki ± 0%  -34.02% (p=0.002 n=6)
geomean                         53.51Ki        43.02Ki       -19.61%                 33.81Ki       -36.83%

                              │  main.txt   │            this-pr.txt             │           pr-15150v2.txt           │
                              │  allocs/op  │  allocs/op   vs base               │  allocs/op   vs base               │
Parse omtestdata.txt omtext      317.0 ± 0%    264.0 ± 0%  -16.72% (p=0.002 n=6)    211.0 ± 0%  -33.44% (p=0.002 n=6)
Parse promtestdata.txt omtext   3.102k ± 0%   2.692k ± 0%  -13.22% (p=0.002 n=6)   2.282k ± 0%  -26.43% (p=0.002 n=6)
geomean                          991.6         843.0       -14.99%                  693.9       -30.02%
```

[mem profile](https://pprof.me/485d947268c2cc87ab798c9f4654fdd8/?profileType=profile%3Aalloc_objects%3Acount%3Aspace%3Abytes)