@Maniktherana 2024-10-11
thanos snapped `p.isCreatedSeries`

<img width="571" alt="Screenshot 2024-10-11 at 10 42 26 PM" src="https://github.com/user-attachments/assets/ce9df4ed-3b1c-44a3-b7b8-9761ae7b1d63">